TT#14008 don't try to offer RTCP after previous mux
If an offer is going to a side that is already known to support
rtcp-mux, don't try to offer a fallback RTCP port. Do this by pretending
that rtcp-mux=require is set, which leads to the correct behaviour. This
is needed because the RTCP fallback port might already be closed, so
trying to include it in the offer would require opening a new set of
ports, which is undesirable.

closes #1494

Change-Id: I550bec08379c799cb7dd090a70d090ae47462467