MT#55283 single-threaded lws_write()
Avoid calling lws_write() from threads other than the service thread, as
this might not be thread-safe. Instead store the values used for the
HTTP response headers in the websocket_output, then trigger a "writable"
callback, and finally do all the lws_write() calls from the service
thread.

Reported in #1624

Change-Id: Ifcb050193044e5543f750a12fb44f5e16d4c0a08