MT#55283 fix DTMF send scheduling
Don't try to interpret or schedule DTMF packets if we're just doing RTP
passthrough.

Without tracking the seqencing and timestamping of the primary media
stream, we cannot actually schedule DTMF packets properly. Use the
presence of a defined output sequence number to make this determination.

The other case that uses a predefined output sequence number is when
media is coming from a pre-coded media player stream. The same thing
applies: These packets are already scheduled, so we can skip trying to
schedule them again.

closes #1638

Change-Id: I2d2c20e731f20a3136ca3eee37f0b521062cdb11